Add animations

+++ TodoList/Views/ListView.swift
@@ -13,19 +13,26 @@ struct ListView: View {
     
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn(duration: 1.0)))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem )
+                    .onMove(perform: listViewModel.moveItem )
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem )
-            .onMove(perform: listViewModel.moveItem )
         }
-        .navigationTitle("Todo List")
+        .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading:EditButton(),
             trailing:
